fix: Resolve clippy warnings and compilation errors across workspace
This commit fixes hundreds of clippy warnings and compilation errors to
meet the zero-tolerance checklist requirements:

**Core Fixes:**
- Fix unused imports (chrono::Utc, PdfParams, JobStatus, redis_helpers)
- Fix dead code warnings with #[allow(dead_code)] on test helpers
- Fix type mismatches (u32→usize, u64→f64 casts)
- Fix partial move errors using .as_ref()
- Add missing tracing macro imports (debug, error, info, warn)

**Error Handling:**
- Add From<ResourceManagerError> for ApiError conversion
- Fix ?-operator error conversions

**Method Signature Fixes:**
- Fix AppState::new signature (removed metrics parameter)
- Fix record_worker_metrics calls (3 args instead of 4)
- Fix record_worker_pool_stats mapping
- Fix record_worker_job_submitted argument

**Test Infrastructure:**
- Fix Redis test helper type errors (query_async signatures)
- Fix PostgreSQL test helper unused methods
- Fix outbox_publisher_tests imports
- Mark test helpers as #[allow(dead_code)]
- Fix clippy::default_constructed_unit_structs

**Code Quality:**
- Replace .len() > 0 with !.is_empty()
- Fix clippy::unnecessary_literal_unwrap
- Remove duplicated #[cfg] attributes
- Fix empty line after doc comments
- Add #[allow(deprecated)] for metrics migration

**Modified Crates:**
- riptide-api: 24 files
- riptide-persistence: 6 files
- riptide-cache: 1 file
- riptide-facade: 2 files
- riptide-browser: 1 file

All library code now compiles with zero warnings under -D warnings.
Test fixes are in progress.

+++ crates/riptide-api/src/handlers/search.rs
@@ -23,7 +23,12 @@ pub async fn search(State(state): State<AppState>, Query(params): Query<SearchQu
     };
 
     match facade
-        .search_validated(&params.q, params.limit, &params.country, &params.language)
+        .search_validated(
+            &params.q,
+            params.limit as usize,
+            &params.country,
+            &params.language,
+        )
         .await
     {
         Ok(hits) => build_response(params.q, hits, start).into_response(),
